init login event set timer

+++ bebas_api/src/main/java/com/fadlurahmanf/bebas_api/data/dto/auth/LoginResponse.kt
@@ -3,6 +3,6 @@ package com.fadlurahmanf.bebas_api.data.dto.auth
 data class LoginResponse(
     var accessToken: String? = null,
     var refreshToken: String? = null,
-    var expiresIn: Int? = null,
-    var refreshExpiresIn: Int? = null
+    var expiresIn: Long? = null,
+    var refreshExpiresIn: Long? = null
 )
